add-register-login-feature

+++ index.php
@@ -3,6 +3,26 @@
 include "bootstrap/init.php";
 
 
+if(isset($_GET['logout'])){
+
+    logedOut();
+}
+
+
+
+
+if(!isLoggedIn()){
+
+    header("Location:http://task-manager.test/auth.php");
+    die();
+}
+
+
+
+
+
+
+
 if( isset($_GET['delete_folder']) and is_numeric($_GET['delete_folder'])){
 
     deleteFolder($_GET['delete_folder']);
